ai_math_tutor/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
bcrypt==4.0.1
PyJWT==2.6.0
redis==4.5.1
flask-cors==3.0.10
flask-sqlalchemy==3.0.3 